# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SDR CRM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces/{workspaceId}/leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads in workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: stage
          in: query
          required: false
          schema:
            type: string
        - name: assigneeId
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a lead
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /workspaces/{workspaceId}/leads/{leadId}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get lead details
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: leadId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lead details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: leadId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadBody"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
        "422":
          description: Stage transition validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StageValidationError"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: leadId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── Campaigns ────────────────────────────────────────────────────────────
...
  /workspaces/{workspaceId}/funnel-stages:
    get:
      operationId: listFunnelStages
      tags: [funnel]
      summary: List funnel stages for workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Funnel stages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FunnelStage"
    post:
      operationId: createFunnelStage
      tags: [funnel]
      summary: Create a funnel stage
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFunnelStageBody"
      responses:
        "201":
          description: Created stage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FunnelStage"
  /workspaces/{workspaceId}/funnel-stages/{stageId}:
    patch:
      operationId: updateFunnelStage
      tags: [funnel]
      summary: Update a funnel stage
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: stageId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFunnelStageBody"
      responses:
        "200":
          description: Updated stage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FunnelStage"
    delete:
      operationId: deleteFunnelStage
      tags: [funnel]
      summary: Delete a funnel stage
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: stageId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── Custom Fields ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Lead:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        stage:
          type: string
        assigneeId:
          type: ["string", "null"]
        customFieldValues:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - workspaceId
        - name
        - stage
        - createdAt
        - updatedAt
    CreateLeadBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        stage:
          type: string
        assigneeId:
          type: ["string", "null"]
        customFieldValues:
          type: object
          additionalProperties: true
      required:
        - name
    UpdateLeadBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        stage:
          type: string
        assigneeId:
          type: ["string", "null"]
        customFieldValues:
          type: object
          additionalProperties: true
    StageValidationError:
      type: object
      properties:
        error:
          type: string
        missingFields:
          type: array
          items:
            type: string
      required:
        - error
        - missingFields
...
    FunnelStage:
      type: object
      properties:
        id:
          type: string
        workspaceId:
          type: string
        name:
          type: string
        slug:
          type: string
        order:
          type: integer
        color:
          type: ["string", "null"]
        requiredFields:
          type: array
          items:
            type: string
        isDefault:
          type: boolean
      required:
        - id
        - workspaceId
        - name
        - slug
        - order
        - isDefault
    CreateFunnelStageBody:
      type: object
      properties:
        name:
          type: string
        color:
          type: ["string", "null"]
        order:
          type: integer
        requiredFields:
          type: array
          items:
            type: string
      required:
        - name
    UpdateFunnelStageBody:
      type: object
      properties:
        name:
          type: string
        color:
          type: ["string", "null"]
        order:
          type: integer
        requiredFields:
          type: array
          items:
            type: string
...